requirements.txt
fastapi
uvicorn
langchain
marqo
llama-cpp-python
python-multipart